requirements.txt
flask
google-cloud-storage
google-cloud-speech
google-genai
google-cloud-secret-manager
python-dotenv
pydub
requests
google-auth
